Change all `Cocoa` references to `AppKit`
It turns out that "Cocoa" refers to the overall API, *including* "AppKit" which is the actual UI layer. By being precise in our naming here, if "Cocoa" or macOS ever move on to a new UI framework, maybe "Cocoa" will live on, but with the new UI libraries. We could support both, but not if both were called "Cocoa".